refactor: 폴더명 life_record -> liferecord로 변경
