+++ src/main/java/com/grepp/teamnotfound/app/controller/api/liferecord/payload/LifeRecordData.java
@@ -1,6 +1,6 @@
-package com.grepp.teamnotfound.app.controller.api.life_record.payload;
+package com.grepp.teamnotfound.app.controller.api.liferecord.payload;
 
-import com.grepp.teamnotfound.app.model.life_record.entity.LifeRecord;
+import com.grepp.teamnotfound.app.model.liferecord.entity.LifeRecord;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.util.List;
