+++ src/main/java/com/grepp/teamnotfound/app/controller/api/liferecord/payload/PageInfo.java
@@ -1,6 +1,6 @@
-package com.grepp.teamnotfound.app.controller.api.life_record.payload;
+package com.grepp.teamnotfound.app.controller.api.liferecord.payload;
 
-import com.grepp.teamnotfound.app.model.life_record.dto.LifeRecordListDto;
+import com.grepp.teamnotfound.app.model.liferecord.dto.LifeRecordListDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
