+++ src/main/java/com/grepp/teamnotfound/app/model/dashboard/DashboardService.java
@@ -7,14 +7,12 @@ import com.grepp.teamnotfound.app.model.dashboard.dto.FeedingDashboardDto;
 import com.grepp.teamnotfound.app.model.dashboard.dto.SleepingDashboardDto;
 import com.grepp.teamnotfound.app.model.dashboard.dto.WalkingDashboardDto;
 import com.grepp.teamnotfound.app.model.dashboard.dto.WeightDashboardDto;
-import com.grepp.teamnotfound.app.model.life_record.LifeRecordService;
-import com.grepp.teamnotfound.app.model.life_record.entity.LifeRecord;
+import com.grepp.teamnotfound.app.model.liferecord.LifeRecordService;
+import com.grepp.teamnotfound.app.model.liferecord.entity.LifeRecord;
 import com.grepp.teamnotfound.app.model.pet.PetService;
 import com.grepp.teamnotfound.app.model.pet.dto.PetDto;
 import com.grepp.teamnotfound.app.model.pet.entity.Pet;
 import com.grepp.teamnotfound.app.model.recommend.DailyRecommendService;
-import com.grepp.teamnotfound.app.model.recommend.RecommendService;
-import com.grepp.teamnotfound.app.model.recommend.repository.DailyRecommendRepository;
 import com.grepp.teamnotfound.app.model.schedule.dto.ScheduleDto;
 import com.grepp.teamnotfound.app.model.schedule.entity.Schedule;
 import com.grepp.teamnotfound.app.model.schedule.repository.ScheduleRepository;
