+++ src/main/java/com/grepp/teamnotfound/app/model/liferecord/dto/LifeRecordDto.java
@@ -1,6 +1,6 @@
-package com.grepp.teamnotfound.app.model.life_record.dto;
+package com.grepp.teamnotfound.app.model.liferecord.dto;
 
-import com.grepp.teamnotfound.app.controller.api.life_record.payload.LifeRecordData;
+import com.grepp.teamnotfound.app.controller.api.liferecord.payload.LifeRecordData;
 import com.grepp.teamnotfound.app.model.structured_data.dto.FeedingDto;
 import com.grepp.teamnotfound.app.model.structured_data.dto.WalkingDto;
 import java.time.LocalDate;
