+++ src/main/java/com/grepp/teamnotfound/app/model/liferecord/repository/LifeRecordRepositoryCustom.java
@@ -1,7 +1,7 @@
-package com.grepp.teamnotfound.app.model.life_record.repository;
+package com.grepp.teamnotfound.app.model.liferecord.repository;
 
-import com.grepp.teamnotfound.app.controller.api.life_record.payload.LifeRecordListRequest;
-import com.grepp.teamnotfound.app.model.life_record.dto.LifeRecordListDto;
+import com.grepp.teamnotfound.app.controller.api.liferecord.payload.LifeRecordListRequest;
+import com.grepp.teamnotfound.app.model.liferecord.dto.LifeRecordListDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
