+++ src/main/java/com/grepp/teamnotfound/app/model/liferecord/repository/LifeRecordRepositoryImpl.java
@@ -1,9 +1,9 @@
-package com.grepp.teamnotfound.app.model.life_record.repository;
+package com.grepp.teamnotfound.app.model.liferecord.repository;
 
-import com.grepp.teamnotfound.app.controller.api.life_record.payload.LifeRecordListRequest;
-import com.grepp.teamnotfound.app.model.life_record.dto.LifeRecordListDto;
-import com.grepp.teamnotfound.app.model.life_record.entity.LifeRecord;
-import com.grepp.teamnotfound.app.model.life_record.entity.QLifeRecord;
+import com.grepp.teamnotfound.app.controller.api.liferecord.payload.LifeRecordListRequest;
+import com.grepp.teamnotfound.app.model.liferecord.dto.LifeRecordListDto;
+import com.grepp.teamnotfound.app.model.liferecord.entity.LifeRecord;
+import com.grepp.teamnotfound.app.model.liferecord.entity.QLifeRecord;
 import com.grepp.teamnotfound.app.model.pet.entity.QPet;
 import com.grepp.teamnotfound.app.model.pet.entity.QPetImg;
 import com.querydsl.core.types.dsl.BooleanExpression;
