+++ src/main/java/com/grepp/teamnotfound/app/model/recommend/RecommendService.java
@@ -6,8 +6,8 @@ import com.grepp.teamnotfound.app.model.recommend.dto.LifeRecordAvgDto;
 import com.grepp.teamnotfound.app.model.recommend.dto.LifeRecordListDto;
 import com.grepp.teamnotfound.app.model.recommend.dto.PetInfoDto;
 import com.grepp.teamnotfound.app.model.recommend.dto.RecommendCheckDto;
-import com.grepp.teamnotfound.app.model.life_record.entity.LifeRecord;
-import com.grepp.teamnotfound.app.model.life_record.repository.LifeRecordRepository;
+import com.grepp.teamnotfound.app.model.liferecord.entity.LifeRecord;
+import com.grepp.teamnotfound.app.model.liferecord.repository.LifeRecordRepository;
 import com.grepp.teamnotfound.app.model.pet.entity.Pet;
 import com.grepp.teamnotfound.app.model.recommend.dto.RecommendStateDto;
 import com.grepp.teamnotfound.app.model.recommend.entity.Recommend;
