+++ src/main/java/com/grepp/teamnotfound/app/model/structured_data/entity/Walking.java
@@ -1,6 +1,6 @@
 package com.grepp.teamnotfound.app.model.structured_data.entity;
 
-import com.grepp.teamnotfound.app.model.life_record.entity.LifeRecord;
+import com.grepp.teamnotfound.app.model.liferecord.entity.LifeRecord;
 import com.grepp.teamnotfound.app.model.structured_data.dto.WalkingDto;
 import com.grepp.teamnotfound.infra.entity.BaseEntity;
 import jakarta.persistence.Column;
